refactor : Line Interface 작업
