+++ src/main/java/nextstep/ladder/domain/lines/line/Line.java
@@ -1,7 +1,7 @@
 package nextstep.ladder.domain.lines.line;
 
 import java.util.List;
-import nextstep.ladder.domain.lines.Point;
+import nextstep.ladder.domain.lines.point.Point;
 
 public interface Line {
