+++ src/main/java/nextstep/ladder/domain/lines/line/impl/LineImpl.java
@@ -3,8 +3,8 @@ package nextstep.ladder.domain.lines.line.impl;
 import java.util.List;
 import java.util.stream.IntStream;
 import nextstep.ladder.domain.ladderConnectOrder.LadderConnectOrder;
-import nextstep.ladder.domain.lines.Point;
 import nextstep.ladder.domain.lines.line.Line;
+import nextstep.ladder.domain.lines.point.Point;
 import nextstep.ladder.error.exception.LineNoConsecutiveTrues;
 
 public class LineImpl implements Line {
